+++ cmd/template/cluster/provider/common.go
@@ -50,10 +50,10 @@ type MachineConfig struct {
 }
 
 type OpenStackConfig struct {
-	Cloud             string
-	CloudConfig       string
-	DNSNameservers    []string
-	EnableOIDC        bool
+	Cloud          string
+	CloudConfig    string
+	DNSNameservers []string
+
 	ExternalNetworkID string
 	NodeCIDR          string
 	NetworkName       string
@@ -85,10 +85,19 @@ type ClusterConfig struct {
 	Labels            map[string]string
 	Namespace         string
 	PodsCIDR          string
+	OIDC              OIDC
 
 	App       AppConfig
 	AWS       AWSConfig
 	OpenStack OpenStackConfig
+}
+
+type OIDC struct {
+	IssuerURL     string
+	CAFile        string
+	ClientID      string
+	UsernameClaim string
+	GroupsClaim   string
 }
 
 type templateConfig struct {